+++ .cspell.json
@@ -282,7 +282,6 @@
 		"setwork",
 		"shadcn",
 		"shandow",
-		"Srcoll",
 		"signin",
 		"Signle",
 		"signoff",
