+++ apps/web/app/hooks/index.ts
@@ -70,4 +70,4 @@ export * from './integrations/useIntegration';
 export * from './integrations/useIntegrationTenant';
 export * from './integrations/useIntegrationTypes';
 
-export * from './useSrcollListener'
+export * from './useScrollListener';
