--- apps/web/app/hooks/useSrcollListener.ts
@@ -1,22 +1,0 @@
-"use client"
-import React from 'react'
-
-export function useSrcollListener() {
-    const [scrolling, setScrolling] = React.useState(false);
-    React.useEffect(() => {
-        const handleScroll = () => {
-            if (window.scrollY > 100) {
-                setScrolling(true);
-            } else {
-                setScrolling(false);
-            }
-        };
-        console.log(window.scrollY)
-        window.addEventListener("scroll", handleScroll);
-        return () => {
-            window.removeEventListener("scroll", handleScroll);
-        };
-    }, []);
-
-    return { scrolling }
-}
